+++ backend-sample/src/main/scala/ru/tinkoff/backendacademy/petapp/repository/pet/LoggingPetRepository.scala
@@ -4,7 +4,7 @@ import zio.ZIO
 import zio.logging.Logging
 
 class LoggingPetRepository(logging: Logging, delegate: PetRepository) extends PetRepository {
-  override def findPetName(petId: Int): ZIO[Any, String, String] = (
+  override def findPetName(petId: Long): ZIO[Any, String, String] = (
     for {
       _   <- Logging.info(s"Looking for pet $petId")
       res <- delegate.findPetName(petId)
