+++ backend-sample/src/main/scala/ru/tinkoff/backendacademy/petapp/repository/pet/SinglePetRepository.scala
@@ -3,7 +3,7 @@ package ru.tinkoff.backendacademy.petapp.repository.pet
 import zio.{IO, ZIO}
 
 object SinglePetRepository extends PetRepository {
-  override def findPetName(petId: Int): ZIO[Any, String, String] =
+  override def findPetName(petId: Long): ZIO[Any, String, String] =
     if (petId == 35) {
       ZIO.succeed("Tapirus terrestris")
     } else {
